Create the smart contract for the Dynamic Consent Management System (DCMS) tailored for people with disabilities (PWDs)
This simple code snippet highlights the use of special and additional functions that should demonstrate the needs of data subjects with some disabilities and need attention. Since these data subjects would also like to participate in sharing their data for research-related assignments, and maybe they want to sell their data. So, we stressed to use of blockchains to have intrinsic features related to the privacy of such people with disabilities (PWDS).